refactor(runtime): split provider lifecycle responsibilities
